Fe inst/junseok

See merge request s10-final/S10P31A604!286

+++ goski_instructor/lib/data/model/student_info.dart
@@ -38,7 +38,7 @@ class StudentInfoResponse {
       age: json['age'],
       height: json['height'],
       weight: json['weight'],
-      footSize: json['footSize'],
+      footSize: json['footSize'] as int,
     );
   }
 }
